+++ hnote/components/molecules/FormItems/SubtaskItem.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const Wrapper = styled.View`
+    display: flex;
+    flex-direction: row;
+`;
+
+interface ISubtaskItem {}
+
+const SubtaskItem = (props) => {};